Arreglos en las alertas

+++ src/app/vistas/favoritos/favoritos.component.ts
@@ -2,6 +2,7 @@ import { ChangeDetectorRef, Component } from '@angular/core';
 import { DarkModeService } from 'angular-dark-mode';
 import { Gasolinera } from 'src/app/clases/gasolinera';
 import { FavoritosService } from 'src/app/servicios/favoritos.service';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-favoritos',
@@ -36,14 +37,38 @@ export class FavoritosComponent {
   }
 
   eliminar(gasolinera: Gasolinera){
-    this.favoritosService.deleteFavoritos(gasolinera);
-    const favoritosString: string | null = localStorage.getItem("favoritos");
-    if (favoritosString !== null) {
-      this.gasolinerasFav = JSON.parse(favoritosString);
-    }
-    if(!this.gasolinerasFav.length){
-      this.datosCargados = false;
-    }
+    Swal.fire({
+      title: `Eliminar gasolinera ${gasolinera.rotulo}`,
+      icon: 'question',
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Si"
+    }).then((result) =>{
+      if(result.isConfirmed){
+        this.favoritosService.deleteFavoritos(gasolinera);
+        const favoritosString: string | null = localStorage.getItem("favoritos");
+        if (favoritosString !== null) {
+          this.gasolinerasFav = JSON.parse(favoritosString);
+        }
+        if(!this.gasolinerasFav.length){
+          this.datosCargados = false;
+        }
+        Swal.fire({
+          title: "Eliminado",
+          icon: "success",
+          showConfirmButton: false,
+          timer: 1100
+        });
+      }else{
+        Swal.fire({
+          title: "No eliminado",
+          icon: "info",
+          showConfirmButton: false,
+          timer: 1100
+        });
+      }
+    })
   }
 
 }
